+++ web_note/web_note_test/__init__.py
@@ -2,3 +2,4 @@
 # © initOS GmbH 2014
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from . import web_note
+from . import res_partner
