+++ web_note/web_note_test/__openerp__.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# © initOS GmbH 2014
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+{
+    'name': 'Web Note Test',
+    'version': '7.0.1.0.0',
+    'author': 'initOS GmbH',
+    'category': '',
+    'description':
+    """
+This module is example of the use of the web_note module.
+    """,
+    'website': 'http://www.initos.com',
+    'license': 'AGPL-3',
+    'images': [],
+    'depends': [
+        "base",
+        "sale",
+        "web_note",
+    ],
+    'data': [
+        "res_partner_view.xml",
+    ],
+    'active': False,
+    'installable': True
+}
