+++ web_note/web_note_test/res_partner_view.xml
