+++ web_note/web_note_view.xml
